References #478. Move forced-native functions to a native/ folder instead of in util/.
