+++ src/api/emit.js
@@ -1,5 +1,5 @@
-import utilCreateElement from '../util/create-element';
-import utilCreateEvent from '../util/create-event';
+import createElement from '../native/create-element';
+import createEvent from '../native/create-event';
 import utilElementContains from '../util/element-contains';
 
 var CustomEvent = (function (CustomEvent) {
@@ -21,8 +21,8 @@ function dispatch (elem, cEvent) {
 }
 
 var hasBubbleOnDetachedElements = (function () {
-  var parent = utilCreateElement('div');
-  var child = utilCreateElement('div');
+  var parent = createElement('div');
+  var child = createElement('div');
   var hasBubbleOnDetachedElements = false;
   parent.appendChild(child);
   parent.addEventListener('test', () => hasBubbleOnDetachedElements = true);
@@ -35,7 +35,7 @@ function createCustomEvent (name, opts = {}) {
     return new CustomEvent(name, opts);
   }
 
-  var e = utilCreateEvent('CustomEvent');
+  var e = createEvent('CustomEvent');
   e.initCustomEvent(name, opts.bubbles, opts.cancelable, opts.detail);
   return e;
 }
