+++ src/native/create-document-fragment.js
