+++ src/native/register-element.js
